+++ packages/transaction-controller/src/TransactionController.ts
@@ -1,5 +1,6 @@
 import type { TypedTransaction } from '@ethereumjs/tx';
 import type {
+  AccountsController,
   AccountsControllerGetSelectedAccountAction,
   AccountsControllerGetStateAction,
 } from '@metamask/accounts-controller';
@@ -120,6 +121,7 @@ import type {
   AddTransactionOptions,
   PublishHookResult,
   GetGasFeeTokensRequest,
+  InternalAccount,
 } from './types';
 import {
   GasFeeEstimateLevel,
@@ -551,6 +553,12 @@ export type TransactionControllerOptions = {
       transactionMeta: TransactionMeta,
     ) => (TransactionMeta | undefined)[];
 
+    /**
+     * Callback to determine whether timeout checking should be enabled for a transaction.
+     * Return false to disable timeout for the transaction.
+     */
+    isTimeoutEnabled?: (transactionMeta: TransactionMeta) => boolean;
+
     /** Alternate logic to publish a transaction. */
     publish?: (
       transactionMeta: TransactionMeta,
@@ -886,6 +894,8 @@ export class TransactionController extends BaseController<
 
   readonly #isSwapsDisabled: boolean;
 
+  readonly #isTimeoutEnabled: (transactionMeta: TransactionMeta) => boolean;
+
   readonly #layer1GasFeeFlows: Layer1GasFeeFlow[];
 
   readonly #methodDataHelper: MethodDataHelper;
@@ -969,44 +979,61 @@ export class TransactionController extends BaseController<
 
     this.messenger = messenger;
 
-    this.#afterAdd = hooks?.afterAdd ?? (() => Promise.resolve({}));
-    this.#afterSign = hooks?.afterSign ?? (() => true);
-    this.#afterSimulate = hooks?.afterSimulate ?? (() => Promise.resolve({}));
+    this.#afterAdd =
+      hooks?.afterAdd ?? ((): ReturnType<AfterAddHook> => Promise.resolve({}));
+    this.#afterSign = hooks?.afterSign ?? ((): boolean => true);
+    this.#afterSimulate =
+      hooks?.afterSimulate ??
+      ((): ReturnType<AfterSimulateHook> => Promise.resolve({}));
     this.#beforeCheckPendingTransaction =
       /* istanbul ignore next */
-      hooks?.beforeCheckPendingTransaction ?? (() => Promise.resolve(true));
-    this.#beforePublish = hooks?.beforePublish ?? (() => Promise.resolve(true));
-    this.#beforeSign = hooks?.beforeSign ?? (() => Promise.resolve({}));
+      hooks?.beforeCheckPendingTransaction ??
+      ((): Promise<boolean> => Promise.resolve(true));
+    this.#beforePublish =
+      hooks?.beforePublish ?? ((): Promise<boolean> => Promise.resolve(true));
+    this.#beforeSign =
+      hooks?.beforeSign ??
+      ((): ReturnType<BeforeSignHook> => Promise.resolve({}));
     this.#getAdditionalSignArguments =
-      hooks?.getAdditionalSignArguments ?? (() => []);
+      hooks?.getAdditionalSignArguments ??
+      ((): (TransactionMeta | undefined)[] => []);
     this.#getCurrentAccountEIP1559Compatibility =
-      getCurrentAccountEIP1559Compatibility ?? (() => Promise.resolve(true));
+      getCurrentAccountEIP1559Compatibility ??
+      ((): Promise<boolean> => Promise.resolve(true));
     this.#getCurrentNetworkEIP1559Compatibility =
       getCurrentNetworkEIP1559Compatibility;
     this.#getExternalPendingTransactions =
-      getExternalPendingTransactions ?? (() => []);
+      getExternalPendingTransactions ?? ((): NonceTrackerTransaction[] => []);
     this.#getGasFeeEstimates =
-      getGasFeeEstimates || (() => Promise.resolve({} as GasFeeState));
+      getGasFeeEstimates ??
+      ((): Promise<GasFeeState> => Promise.resolve({} as GasFeeState));
     this.#getNetworkState = getNetworkState;
     this.#getPermittedAccounts = getPermittedAccounts;
-    this.#getSavedGasFees = getSavedGasFees ?? ((_chainId) => undefined);
+    this.#getSavedGasFees =
+      getSavedGasFees ?? ((_chainId): SavedGasFees | undefined => undefined);
     this.#getSimulationConfig =
-      getSimulationConfig ?? (() => Promise.resolve({}));
+      getSimulationConfig ??
+      ((): ReturnType<GetSimulationConfig> => Promise.resolve({}));
     this.#incomingTransactionOptions = incomingTransactions;
     this.#isAutomaticGasFeeUpdateEnabled =
-      isAutomaticGasFeeUpdateEnabled ?? ((_txMeta: TransactionMeta) => false);
+      isAutomaticGasFeeUpdateEnabled ??
+      ((_txMeta: TransactionMeta): boolean => false);
     this.#isEIP7702GasFeeTokensEnabled =
-      isEIP7702GasFeeTokensEnabled ?? (() => Promise.resolve(false));
+      isEIP7702GasFeeTokensEnabled ??
+      ((): Promise<boolean> => Promise.resolve(false));
     this.#isFirstTimeInteractionEnabled =
-      isFirstTimeInteractionEnabled ?? (() => true);
+      isFirstTimeInteractionEnabled ?? ((): boolean => true);
     this.#isHistoryDisabled = disableHistory ?? false;
     this.#isSendFlowHistoryDisabled = disableSendFlowHistory ?? false;
-    this.#isSimulationEnabled = isSimulationEnabled ?? (() => true);
+    this.#isSimulationEnabled = isSimulationEnabled ?? ((): boolean => true);
     this.#isSwapsDisabled = disableSwaps ?? false;
+    this.#isTimeoutEnabled = hooks?.isTimeoutEnabled ?? ((): boolean => true);
     this.#pendingTransactionOptions = pendingTransactions;
     this.#publicKeyEIP7702 = publicKeyEIP7702;
     this.#publish =
-      hooks?.publish ?? (() => Promise.resolve({ transactionHash: undefined }));
+      hooks?.publish ??
+      ((): Promise<{ transactionHash?: string }> =>
+        Promise.resolve({ transactionHash: undefined }));
     this.#publishBatchHook = hooks?.publishBatch;
     this.#securityProviderRequest = securityProviderRequest;
     this.#sign = sign;
@@ -1014,7 +1041,7 @@ export class TransactionController extends BaseController<
     this.#trace = trace ?? (((_request, fn) => fn?.()) as TraceCallback);
     this.#transactionHistoryLimit = transactionHistoryLimit;
 
-    const findNetworkClientIdByChainId = (chainId: Hex) => {
+    const findNetworkClientIdByChainId = (chainId: Hex): string => {
       return this.messenger.call(
         `NetworkController:findNetworkClientIdByChainId`,
         chainId,
@@ -1035,7 +1062,7 @@ export class TransactionController extends BaseController<
       createNonceTracker: this.#createNonceTracker.bind(this),
       createPendingTransactionTracker:
         this.#createPendingTransactionTracker.bind(this),
-      onNetworkStateChange: (listener) => {
+      onNetworkStateChange: (listener): void => {
         this.messenger.subscribe('NetworkController:stateChange', listener);
       },
     });
@@ -1048,12 +1075,14 @@ export class TransactionController extends BaseController<
       findNetworkClientIdByChainId,
       gasFeeFlows: this.#gasFeeFlows,
       getGasFeeControllerEstimates: this.#getGasFeeEstimates,
-      getProvider: (networkClientId) => this.#getProvider({ networkClientId }),
-      getTransactions: () => this.state.transactions,
-      getTransactionBatches: () => this.state.transactionBatches,
+      getProvider: (networkClientId): Provider =>
+        this.#getProvider({ networkClientId }),
+      getTransactions: (): TransactionMeta[] => this.state.transactions,
+      getTransactionBatches: (): TransactionBatchMeta[] =>
+        this.state.transactionBatches,
       layer1GasFeeFlows: this.#layer1GasFeeFlows,
       messenger: this.messenger,
-      onStateChange: (listener) => {
+      onStateChange: (listener): void => {
         this.messenger.subscribe('TransactionController:stateChange', listener);
       },
     });
@@ -1069,8 +1098,9 @@ export class TransactionController extends BaseController<
     );
 
     this.#methodDataHelper = new MethodDataHelper({
-      getProvider: (networkClientId) => this.#getProvider({ networkClientId }),
-      getState: () => this.state.methodData,
+      getProvider: (networkClientId): Provider =>
+        this.#getProvider({ networkClientId }),
+      getState: (): Record<string, MethodData> => this.state.methodData,
     });
 
     this.#methodDataHelper.hub.on(
@@ -1084,8 +1114,10 @@ export class TransactionController extends BaseController<
 
     this.#incomingTransactionHelper = new IncomingTransactionHelper({
       client: this.#incomingTransactionOptions.client,
-      getCurrentAccount: () => this.#getSelectedAccount(),
-      getLocalTransactions: () => this.state.transactions,
+      getCurrentAccount: (): ReturnType<
+        AccountsController['getSelectedAccount']
+      > => this.#getSelectedAccount(),
+      getLocalTransactions: (): TransactionMeta[] => this.state.transactions,
       includeTokenTransfers:
         this.#incomingTransactionOptions.includeTokenTransfers,
       isEnabled: this.#incomingTransactionOptions.isEnabled,
@@ -1106,16 +1138,17 @@ export class TransactionController extends BaseController<
       this.#checkForPendingTransactionAndStartPolling,
     );
 
+    // eslint-disable-next-line no-new
     new ResimulateHelper({
       simulateTransaction: this.#updateSimulationData.bind(this),
-      onTransactionsUpdate: (listener) => {
+      onTransactionsUpdate: (listener): void => {
         this.messenger.subscribe(
           'TransactionController:stateChange',
           listener,
           (controllerState) => controllerState.transactions,
         );
       },
-      getTransactions: () => this.state.transactions,
+      getTransactions: (): TransactionMeta[] => this.state.transactions,
     });
 
     this.#onBootCleanup();
@@ -1126,7 +1159,7 @@ export class TransactionController extends BaseController<
   /**
    * Stops polling and removes listeners to prepare the controller for garbage collection.
    */
-  destroy() {
+  destroy(): void {
     this.#stopAllTracking();
   }
 
@@ -1245,6 +1278,7 @@ export class TransactionController extends BaseController<
       type,
     } = options;
 
+    // eslint-disable-next-line no-param-reassign
     txParams = normalizeTransactionParams(txParams);
 
     if (!this.#multichainTrackingHelper.has(networkClientId)) {
@@ -1356,6 +1390,7 @@ export class TransactionController extends BaseController<
       updateTransaction(addedTransactionMeta);
     }
 
+    // eslint-disable-next-line no-negated-condition
     if (!skipInitialGasEstimate) {
       await this.#trace(
         { name: 'Estimate Gas Properties', parentContext: traceContext },
@@ -1399,11 +1434,13 @@ export class TransactionController extends BaseController<
           addedTransactionMeta,
           method,
         );
+        // eslint-disable-next-line require-atomic-updates
         addedTransactionMeta.securityProviderResponse =
           securityProviderResponse;
       }
 
       if (!this.#isSendFlowHistoryDisabled) {
+        // eslint-disable-next-line require-atomic-updates
         addedTransactionMeta.sendFlowHistory = sendFlowHistory ?? [];
       }
       // Initial history push
@@ -1442,6 +1479,7 @@ export class TransactionController extends BaseController<
         })
         .catch(noop);
 
+      // eslint-disable-next-line no-negated-condition
       if (requireApproval !== false) {
         this.#updateSimulationData(addedTransactionMeta, {
           traceContext,
@@ -1486,11 +1524,11 @@ export class TransactionController extends BaseController<
     };
   }
 
-  startIncomingTransactionPolling() {
+  startIncomingTransactionPolling(): void {
     this.#incomingTransactionHelper.start();
   }
 
-  stopIncomingTransactionPolling() {
+  stopIncomingTransactionPolling(): void {
     this.#incomingTransactionHelper.stop();
   }
 
@@ -1500,7 +1538,9 @@ export class TransactionController extends BaseController<
    * @param request - Request object.
    * @param request.tags - Additional tags to identify the source of the request.
    */
-  async updateIncomingTransactions({ tags }: { tags?: string[] } = {}) {
+  async updateIncomingTransactions({
+    tags,
+  }: { tags?: string[] } = {}): Promise<void> {
     await this.#incomingTransactionHelper.update({ tags });
   }
 
@@ -1521,7 +1561,7 @@ export class TransactionController extends BaseController<
       estimatedBaseFee,
       actionId,
     }: { estimatedBaseFee?: string; actionId?: string } = {},
-  ) {
+  ): Promise<void> {
     await this.#retryTransaction({
       actionId,
       estimatedBaseFee,
@@ -1565,7 +1605,7 @@ export class TransactionController extends BaseController<
       actionId,
       estimatedBaseFee,
     }: { actionId?: string; estimatedBaseFee?: string } = {},
-  ) {
+  ): Promise<void> {
     await this.#retryTransaction({
       actionId,
       estimatedBaseFee,
@@ -1603,7 +1643,7 @@ export class TransactionController extends BaseController<
     rate: number;
     transactionId: string;
     transactionType: TransactionType;
-  }) {
+  }): Promise<void> {
     // If transaction is found for same action id, do not create a new transaction.
     if (this.#getTransactionWithActionId(actionId)) {
       return;
@@ -1611,6 +1651,7 @@ export class TransactionController extends BaseController<
 
     if (gasValues) {
       // Not good practice to reassign a parameter but temporarily avoiding a larger refactor.
+      // eslint-disable-next-line no-param-reassign
       gasValues = normalizeGasFeeValues(gasValues);
       validateGasValues(gasValues);
     }
@@ -1721,7 +1762,10 @@ export class TransactionController extends BaseController<
     }: {
       ignoreDelegationSignatures?: boolean;
     } = {},
-  ) {
+  ): Promise<{
+    gas: string;
+    simulationFails: TransactionMeta['simulationFails'];
+  }> {
     const ethQuery = this.#getEthQuery({
       networkClientId,
     });
@@ -1751,7 +1795,10 @@ export class TransactionController extends BaseController<
     transaction: TransactionParams,
     multiplier: number,
     networkClientId: NetworkClientId,
-  ) {
+  ): Promise<{
+    gas: string;
+    simulationFails: TransactionMeta['simulationFails'];
+  }> {
     const ethQuery = this.#getEthQuery({
       networkClientId,
     });
@@ -1779,7 +1826,7 @@ export class TransactionController extends BaseController<
    * @param transactionMeta - The new transaction to store in state.
    * @param note - A note or update reason to include in the transaction history.
    */
-  updateTransaction(transactionMeta: TransactionMeta, note: string) {
+  updateTransaction(transactionMeta: TransactionMeta, note: string): void {
     const { id: transactionId } = transactionMeta;
 
     this.#updateTransactionInternal({ transactionId, note }, () => ({
@@ -1796,7 +1843,7 @@ export class TransactionController extends BaseController<
   updateSecurityAlertResponse(
     transactionId: string,
     securityAlertResponse: SecurityAlertResponse,
-  ) {
+  ): void {
     if (!securityAlertResponse) {
       throw new Error(
         'updateSecurityAlertResponse: securityAlertResponse should not be null',
@@ -1831,7 +1878,7 @@ export class TransactionController extends BaseController<
   }: {
     address?: string;
     chainId?: string;
-  } = {}) {
+  } = {}): void {
     if (!chainId && !address) {
       this.update((state) => {
         state.transactions = [];
@@ -1874,7 +1921,7 @@ export class TransactionController extends BaseController<
     transactionMeta: TransactionMeta,
     transactionReceipt: TransactionReceipt,
     baseFeePerGas: Hex,
-  ) {
+  ): Promise<void> {
     // Run validation and add external transaction to state.
     const newTransactionMeta = this.#addExternalTransaction(transactionMeta);
 
@@ -2203,7 +2250,7 @@ export class TransactionController extends BaseController<
       updateType?: boolean;
       value?: string;
     },
-  ) {
+  ): Promise<Readonly<TransactionMeta> | undefined> {
     const transactionMeta = this.#getTransaction(txId);
 
     if (!transactionMeta) {
@@ -2271,7 +2318,7 @@ export class TransactionController extends BaseController<
    * @param transactionId - The ID of the transaction to update.
    * @param isActive - The active state.
    */
-  setTransactionActive(transactionId: string, isActive: boolean) {
+  setTransactionActive(transactionId: string, isActive: boolean): void {
     const transactionMeta = this.#getTransaction(transactionId);
 
     if (!transactionMeta) {
@@ -2348,11 +2395,11 @@ export class TransactionController extends BaseController<
           return this.#signExternalTransaction(txParams.chainId, txParams);
         }),
       );
-    } catch (err) {
-      log('Error while signing transactions with same nonce', err);
+    } catch (error) {
+      log('Error while signing transactions with same nonce', error);
       // Must set transaction to submitted/failed before releasing lock
       // continue with error chain
-      throw err;
+      throw error;
     } finally {
       nonceLock?.releaseLock();
       this.#approvingTransactionIds.delete(initialTxAsSerializedHex);
@@ -2367,7 +2414,9 @@ export class TransactionController extends BaseController<
    *
    * @returns The updated transaction metadata.
    */
-  updateCustodialTransaction(request: UpdateCustodialTransactionRequest) {
+  updateCustodialTransaction(
+    request: UpdateCustodialTransactionRequest,
+  ): TransactionMeta {
     const {
       transactionId,
       errorMessage,
@@ -2488,7 +2537,7 @@ export class TransactionController extends BaseController<
         ? predicate
         : // TODO: Replace `any` with type
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          (v: any) => v === predicate;
+          (value: any): boolean => value === predicate;
     });
 
     const transactionsToFilter = initialList ?? this.state.transactions;
@@ -2665,7 +2714,7 @@ export class TransactionController extends BaseController<
   /**
    * Removes unapproved transactions from state.
    */
-  clearUnapprovedTransactions() {
+  clearUnapprovedTransactions(): void {
     const transactions = this.state.transactions.filter(
       ({ status }) => status !== TransactionStatus.unapproved,
     );
@@ -2680,7 +2729,7 @@ export class TransactionController extends BaseController<
    *
    * @param transactionId - The ID of the transaction to stop signing.
    */
-  abortTransactionSigning(transactionId: string) {
+  abortTransactionSigning(transactionId: string): void {
     const transactionMeta = this.#getTransaction(transactionId);
 
     if (!transactionMeta) {
@@ -2717,7 +2766,7 @@ export class TransactionController extends BaseController<
     transactionId: string;
     transactionIndex: number;
     transactionData: Hex;
-  }) {
+  }): Promise<Hex> {
     log('Updating atomic batch data', {
       transactionId,
       transactionIndex,
@@ -2791,7 +2840,7 @@ export class TransactionController extends BaseController<
   }: {
     transactionId: string;
     batchTransactions: BatchTransactionParams[];
-  }) {
+  }): void {
     log('Updating batch transactions', { transactionId, batchTransactions });
 
     this.#updateTransactionInternal(
@@ -2814,7 +2863,7 @@ export class TransactionController extends BaseController<
   updateSelectedGasFeeToken(
     transactionId: string,
     contractAddress: Hex | undefined,
-  ) {
+  ): void {
     this.#updateTransactionInternal({ transactionId }, (transactionMeta) => {
       const hasMatchingGasFeeToken = transactionMeta.gasFeeTokens?.some(
         (token) =>
@@ -2847,7 +2896,7 @@ export class TransactionController extends BaseController<
     transactionId: string;
     requiredTransactionIds: string[];
     append?: boolean;
-  }) {
+  }): void {
     this.#updateTransactionInternal({ transactionId }, (transactionMeta) => {
       const { requiredTransactionIds: existing } = transactionMeta;
 
@@ -2863,7 +2912,7 @@ export class TransactionController extends BaseController<
    *
    * @param transactionId - The transaction ID.
    */
-  emulateNewTransaction(transactionId: string) {
+  emulateNewTransaction(transactionId: string): void {
     const transactionMeta = this.state.transactions.find(
       (tx) => tx.id === transactionId,
     );
@@ -2889,7 +2938,7 @@ export class TransactionController extends BaseController<
    *
    * @param transactionMeta - Transaction metadata.
    */
-  emulateTransactionUpdate(transactionMeta: TransactionMeta) {
+  emulateTransactionUpdate(transactionMeta: TransactionMeta): void {
     const updatedTransactionMeta = {
       ...transactionMeta,
       txParams: {
@@ -2919,7 +2968,7 @@ export class TransactionController extends BaseController<
     });
   }
 
-  #addMetadata(transactionMeta: TransactionMeta) {
+  #addMetadata(transactionMeta: TransactionMeta): void {
     validateTxParams(transactionMeta.txParams);
     this.update((state) => {
       state.transactions = this.#trimTransactionsForState([
@@ -2932,7 +2981,7 @@ export class TransactionController extends BaseController<
   async #updateGasProperties(
     transactionMeta: TransactionMeta,
     { traceContext }: { traceContext?: TraceContext } = {},
-  ) {
+  ): Promise<void> {
     const isEIP1559Compatible =
       transactionMeta.txParams.type !== TransactionEnvelopeType.legacy &&
       (await this.#getEIP1559Compatibility(transactionMeta.networkClientId));
@@ -2974,12 +3023,12 @@ export class TransactionController extends BaseController<
     );
   }
 
-  #onBootCleanup() {
+  #onBootCleanup(): void {
     this.clearUnapprovedTransactions();
     this.#failIncompleteTransactions();
   }
 
-  #failIncompleteTransactions() {
+  #failIncompleteTransactions(): void {
     const incompleteTransactions = this.state.transactions.filter(
       (transaction) =>
         [TransactionStatus.approved, TransactionStatus.signed].includes(
@@ -3095,10 +3144,12 @@ export class TransactionController extends BaseController<
           const updatedTransactionMeta = this.#getTransaction(
             transactionId,
           ) as TransactionMeta;
-          this.messenger.publish(`${controllerName}:transactionApproved`, {
-            transactionMeta: updatedTransactionMeta,
-            actionId,
-          });
+          if (approvalResult === ApprovalState.Approved) {
+            this.messenger.publish(`${controllerName}:transactionApproved`, {
+              transactionMeta: updatedTransactionMeta,
+              actionId,
+            });
+          }
         }
       } catch (rawError: unknown) {
         const error = rawError as Error & { code?: number; data?: Json };
@@ -3121,24 +3172,26 @@ export class TransactionController extends BaseController<
     const finalMeta = await finishedPromise;
 
     switch (finalMeta?.status) {
-      case TransactionStatus.failed:
+      case TransactionStatus.failed: {
         const error = finalMeta.error as Error;
         resultCallbacks?.error(error);
         throw rpcErrors.internal(error.message);
+      }
 
       case TransactionStatus.submitted:
         resultCallbacks?.success();
         return finalMeta.hash as string;
 
-      default:
+      default: {
         const internalError = rpcErrors.internal(
           `MetaMask Tx Signature: Unknown problem: ${JSON.stringify(
-            finalMeta || transactionId,
+            finalMeta ?? transactionId,
           )}`,
         );
 
         resultCallbacks?.error(internalError);
         throw internalError;
+      }
     }
   }
 
@@ -3157,7 +3210,7 @@ export class TransactionController extends BaseController<
     transactionId: string,
     traceContext?: unknown,
     publishHookOverride?: PublishHook,
-  ) {
+  ): Promise<ApprovalState> {
     let clearApprovingTransactionId: (() => void) | undefined;
     let clearNonceLock: (() => void) | undefined;
 
@@ -3187,7 +3240,7 @@ export class TransactionController extends BaseController<
 
       this.#approvingTransactionIds.add(transactionId);
 
-      clearApprovingTransactionId = () =>
+      clearApprovingTransactionId = (): boolean =>
         this.#approvingTransactionIds.delete(transactionId);
 
       const { networkClientId } = transactionMeta;
@@ -3204,6 +3257,7 @@ export class TransactionController extends BaseController<
 
       clearNonceLock = releaseNonce;
 
+      // eslint-disable-next-line require-atomic-updates
       transactionMeta = this.#updateTransactionInternal(
         {
           transactionId,
@@ -3231,6 +3285,7 @@ export class TransactionController extends BaseController<
         () => this.#signTransaction(transactionMeta),
       );
 
+      // eslint-disable-next-line require-atomic-updates
       transactionMeta = this.#getTransactionOrThrow(transactionId);
 
       if (!(await this.#beforePublish(transactionMeta))) {
@@ -3286,6 +3341,7 @@ export class TransactionController extends BaseController<
         rawTx ?? '0x',
       );
 
+      // eslint-disable-next-line require-atomic-updates
       transactionMeta = this.#updateTransactionInternal(
         {
           transactionId,
@@ -3349,7 +3405,11 @@ export class TransactionController extends BaseController<
    * @param actionId - The actionId passed from UI
    * @param error - The error that caused the rejection.
    */
-  #rejectTransaction(transactionId: string, actionId?: string, error?: Error) {
+  #rejectTransaction(
+    transactionId: string,
+    actionId?: string,
+    error?: Error,
+  ): void {
     const transactionMeta = this.#getTransaction(transactionId);
 
     if (!transactionMeta) {
@@ -3481,7 +3541,7 @@ export class TransactionController extends BaseController<
       'ApprovalController:addRequest',
       {
         id,
-        origin: origin || ORIGIN_METAMASK,
+        origin: origin ?? ORIGIN_METAMASK,
         type,
         requestData,
         expectsResult: true,
@@ -3510,7 +3570,7 @@ export class TransactionController extends BaseController<
     return txMeta;
   }
 
-  #getApprovalId(txMeta: TransactionMeta) {
+  #getApprovalId(txMeta: TransactionMeta): string {
     return String(txMeta.id);
   }
 
@@ -3540,7 +3600,7 @@ export class TransactionController extends BaseController<
   }: {
     chainId?: Hex;
     networkClientId?: NetworkClientId;
-  }) {
+  }): NetworkClientId {
     if (networkClientId) {
       return networkClientId;
     }
@@ -3573,7 +3633,7 @@ export class TransactionController extends BaseController<
     }).provider;
   }
 
-  #onIncomingTransactions(transactions: TransactionMeta[]) {
+  #onIncomingTransactions(transactions: TransactionMeta[]): void {
     if (!transactions.length) {
       return;
     }
@@ -3663,7 +3723,7 @@ export class TransactionController extends BaseController<
    * @param transactionMeta - Nominated external transaction to be added to state.
    * @returns The new transaction.
    */
-  #addExternalTransaction(transactionMeta: TransactionMeta) {
+  #addExternalTransaction(transactionMeta: TransactionMeta): TransactionMeta {
     const { chainId } = transactionMeta;
     const { transactions } = this.state;
     const fromAddress = transactionMeta?.txParams?.from;
@@ -3707,7 +3767,7 @@ export class TransactionController extends BaseController<
    *
    * @param transactionId - Used to identify original transaction.
    */
-  #markNonceDuplicatesDropped(transactionId: string) {
+  #markNonceDuplicatesDropped(transactionId: string): void {
     const transactionMeta = this.#getTransaction(transactionId);
     if (!transactionMeta) {
       return;
@@ -3757,7 +3817,7 @@ export class TransactionController extends BaseController<
    *
    * @param transactionMeta - TransactionMeta of transaction to be marked as dropped.
    */
-  #setTransactionStatusDropped(transactionMeta: TransactionMeta) {
+  #setTransactionStatusDropped(transactionMeta: TransactionMeta): void {
     const updatedTransactionMeta = {
       ...transactionMeta,
       status: TransactionStatus.dropped as const,
@@ -3778,7 +3838,7 @@ export class TransactionController extends BaseController<
    * @param actionId - Unique ID to prevent duplicate requests
    * @returns the filtered transaction
    */
-  #getTransactionWithActionId(actionId?: string) {
+  #getTransactionWithActionId(actionId?: string): TransactionMeta | undefined {
     return this.state.transactions.find(
       (transaction) => actionId && transaction.actionId === actionId,
     );
@@ -3821,7 +3881,9 @@ export class TransactionController extends BaseController<
     return transactionMetaWithRsv;
   }
 
-  async #getEIP1559Compatibility(networkClientId?: NetworkClientId) {
+  async #getEIP1559Compatibility(
+    networkClientId?: NetworkClientId,
+  ): Promise<boolean> {
     const currentNetworkIsEIP1559Compatible =
       await this.#getCurrentNetworkEIP1559Compatibility(networkClientId);
 
@@ -3958,7 +4020,7 @@ export class TransactionController extends BaseController<
     return rawTx;
   }
 
-  #onTransactionStatusChange(transactionMeta: TransactionMeta) {
+  #onTransactionStatusChange(transactionMeta: TransactionMeta): void {
     this.messenger.publish(`${controllerName}:transactionStatusUpdated`, {
       transactionMeta,
     });
@@ -3968,7 +4030,7 @@ export class TransactionController extends BaseController<
     statuses: TransactionStatus[],
     address: string,
     chainId: string,
-  ) {
+  ): NonceTrackerTransaction[] {
     return getAndFormatTransactionsForNonceTracker(
       chainId,
       address,
@@ -3977,7 +4039,7 @@ export class TransactionController extends BaseController<
     );
   }
 
-  #onConfirmedTransaction(transactionMeta: TransactionMeta) {
+  #onConfirmedTransaction(transactionMeta: TransactionMeta): void {
     log('Processing confirmed transaction', transactionMeta.id);
 
     this.#markNonceDuplicatesDropped(transactionMeta.id);
@@ -3996,7 +4058,7 @@ export class TransactionController extends BaseController<
     });
   }
 
-  async #updatePostBalance(transactionMeta: TransactionMeta) {
+  async #updatePostBalance(transactionMeta: TransactionMeta): Promise<void> {
     try {
       const { networkClientId, type } = transactionMeta;
 
@@ -4068,24 +4130,25 @@ export class TransactionController extends BaseController<
 
     const pendingTransactionTracker = new PendingTransactionTracker({
       blockTracker,
-      getChainId: () => chainId,
-      getEthQuery: () => ethQuery,
-      getNetworkClientId: () => networkClientId,
-      getTransactions: () => this.state.transactions,
-      isResubmitEnabled: this.#pendingTransactionOptions.isResubmitEnabled,
-      getGlobalLock: () =>
+      getChainId: (): Hex => chainId,
+      getEthQuery: (): EthQuery => ethQuery,
+      getGlobalLock: (): Promise<() => void> =>
         this.#multichainTrackingHelper.acquireNonceLockForChainIdKey({
           chainId,
         }),
-      messenger: this.messenger,
-      publishTransaction: (_ethQuery, transactionMeta) =>
-        this.#publishTransaction(_ethQuery, transactionMeta, {
-          skipSubmitHistory: true,
-        }),
+      getNetworkClientId: (): NetworkClientId => networkClientId,
+      getTransactions: (): TransactionMeta[] => this.state.transactions,
       hooks: {
         beforeCheckPendingTransaction:
           this.#beforeCheckPendingTransaction.bind(this),
       },
+      isResubmitEnabled: this.#pendingTransactionOptions.isResubmitEnabled,
+      isTimeoutEnabled: this.#isTimeoutEnabled,
+      messenger: this.messenger,
+      publishTransaction: (_ethQuery, transactionMeta): Promise<string> =>
+        this.#publishTransaction(_ethQuery, transactionMeta, {
+          skipSubmitHistory: true,
+        }),
     });
 
     this.#addPendingTransactionTrackerListeners(pendingTransactionTracker);
@@ -4093,17 +4156,17 @@ export class TransactionController extends BaseController<
     return pendingTransactionTracker;
   }
 
-  readonly #checkForPendingTransactionAndStartPolling = () => {
+  readonly #checkForPendingTransactionAndStartPolling = (): void => {
     this.#multichainTrackingHelper.checkForPendingTransactionAndStartPolling();
   };
 
-  #stopAllTracking() {
+  #stopAllTracking(): void {
     this.#multichainTrackingHelper.stopAllTracking();
   }
 
   #addIncomingTransactionHelperListeners(
     incomingTransactionHelper: IncomingTransactionHelper,
-  ) {
+  ): void {
     incomingTransactionHelper.hub.on(
       'transactions',
       this.#onIncomingTransactions.bind(this),
@@ -4112,7 +4175,7 @@ export class TransactionController extends BaseController<
 
   #removePendingTransactionTrackerListeners(
     pendingTransactionTracker: PendingTransactionTracker,
-  ) {
+  ): void {
     pendingTransactionTracker.hub.removeAllListeners('transaction-confirmed');
     pendingTransactionTracker.hub.removeAllListeners('transaction-dropped');
     pendingTransactionTracker.hub.removeAllListeners('transaction-failed');
@@ -4121,7 +4184,7 @@ export class TransactionController extends BaseController<
 
   #addPendingTransactionTrackerListeners(
     pendingTransactionTracker: PendingTransactionTracker,
-  ) {
+  ): void {
     pendingTransactionTracker.hub.on(
       'transaction-confirmed',
       this.#onConfirmedTransaction.bind(this),
@@ -4143,7 +4206,10 @@ export class TransactionController extends BaseController<
     );
   }
 
-  #getNonceTrackerPendingTransactions(chainId: string, address: string) {
+  #getNonceTrackerPendingTransactions(
+    chainId: string,
+    address: string,
+  ): NonceTrackerTransaction[] {
     const standardPendingTransactions = this.#getNonceTrackerTransactions(
       [
         TransactionStatus.approved,
@@ -4186,8 +4252,8 @@ export class TransactionController extends BaseController<
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   #isTransactionAlreadyConfirmedError(error: any): boolean {
     return (
-      error?.message?.includes('nonce too low') ||
-      error?.data?.message?.includes('nonce too low')
+      Boolean(error?.message?.includes('nonce too low')) ||
+      Boolean(error?.data?.message?.includes('nonce too low'))
     );
   }
 
@@ -4293,7 +4359,7 @@ export class TransactionController extends BaseController<
       blockTime?: number;
       traceContext?: TraceContext;
     } = {},
-  ) {
+  ): Promise<void> {
     const {
       chainId,
       id: transactionId,
@@ -4400,7 +4466,7 @@ export class TransactionController extends BaseController<
     gasFeeEstimates?: GasFeeEstimates;
     gasFeeEstimatesLoaded?: boolean;
     layer1GasFee?: Hex;
-  }) {
+  }): void {
     this.#updateTransactionInternal(
       { transactionId, skipHistory: true },
       (txMeta) => {
@@ -4421,7 +4487,7 @@ export class TransactionController extends BaseController<
   }: {
     transactionBatchId: Hex;
     gasFeeEstimates?: GasFeeEstimates;
-  }) {
+  }): void {
     this.#updateTransactionBatch(transactionBatchId, (batch) => {
       return { ...batch, gasFeeEstimates };
     });
@@ -4445,7 +4511,7 @@ export class TransactionController extends BaseController<
     });
   }
 
-  #getSelectedAccount() {
+  #getSelectedAccount(): InternalAccount {
     return this.messenger.call('AccountsController:getSelectedAccount');
   }
 
@@ -4495,7 +4561,7 @@ export class TransactionController extends BaseController<
     });
   }
 
-  async #updateGasEstimate(transactionMeta: TransactionMeta) {
+  async #updateGasEstimate(transactionMeta: TransactionMeta): Promise<void> {
     const { chainId, networkClientId } = transactionMeta;
 
     const isCustomNetwork =
@@ -4572,7 +4638,7 @@ export class TransactionController extends BaseController<
     );
   }
 
-  #deleteTransaction(transactionId: string) {
+  #deleteTransaction(transactionId: string): void {
     this.update((state) => {
       const transactions = state.transactions.filter(
         ({ id }) => id !== transactionId,
@@ -4582,7 +4648,7 @@ export class TransactionController extends BaseController<
     });
   }
 
-  #isRejectError(error: Error & { code?: number }) {
+  #isRejectError(error: Error & { code?: number }): boolean {
     return [
       errorCodes.provider.userRejectedRequest,
       ErrorCode.RejectedUpgrade,
@@ -4593,7 +4659,7 @@ export class TransactionController extends BaseController<
     transactionId: string,
     actionId: string | undefined,
     error: Error & { code?: number; data?: Json },
-  ) {
+  ): void {
     this.#rejectTransaction(transactionId, actionId, error);
 
     if (error.code === errorCodes.provider.userRejectedRequest) {
@@ -4610,7 +4676,7 @@ export class TransactionController extends BaseController<
     transactionMeta: TransactionMeta,
     error: Error,
     actionId?: string,
-  ) {
+  ): void {
     let newTransactionMeta: TransactionMeta;
 
     try {
@@ -4630,8 +4696,8 @@ export class TransactionController extends BaseController<
           ).error = normalizeTxError(error);
         },
       );
-    } catch (err: unknown) {
-      log('Failed to mark transaction as failed', err);
+    } catch (caughtError: unknown) {
+      log('Failed to mark transaction as failed', caughtError);
 
       newTransactionMeta = {
         ...transactionMeta,
@@ -4659,7 +4725,7 @@ export class TransactionController extends BaseController<
     );
   }
 
-  async #runAfterSimulateHook(transactionMeta: TransactionMeta) {
+  async #runAfterSimulateHook(transactionMeta: TransactionMeta): Promise<void> {
     log('Calling afterSimulate hook', transactionMeta);
 
     const { id: transactionId } = transactionMeta;
@@ -4668,7 +4734,7 @@ export class TransactionController extends BaseController<
       transactionMeta,
     });
 
-    const { skipSimulation, updateTransaction } = result || {};
+    const { skipSimulation, updateTransaction } = result ?? {};
 
     if (skipSimulation) {
       this.#skipSimulationTransactionIds.add(transactionId);
@@ -4717,12 +4783,11 @@ export class TransactionController extends BaseController<
 
         ({ transactionHash } = await publishHook(transactionMeta, signedTx));
 
-        if (transactionHash === undefined) {
-          transactionHash = await this.#publishTransaction(ethQuery, {
-            ...transactionMeta,
-            rawTx: signedTx,
-          });
-        }
+        // eslint-disable-next-line require-atomic-updates
+        transactionHash ??= await this.#publishTransaction(ethQuery, {
+          ...transactionMeta,
+          rawTx: signedTx,
+        });
       },
     );
 
@@ -4731,7 +4796,10 @@ export class TransactionController extends BaseController<
     return { transactionHash };
   }
 
-  async #getGasFeeTokens(transaction: TransactionMeta) {
+  async #getGasFeeTokens(transaction: TransactionMeta): Promise<{
+    gasFeeTokens: GasFeeToken[];
+    isGasFeeSponsored: boolean;
+  }> {
     const { chainId } = transaction;
 
     return await getGasFeeTokens({