+++ packages/transaction-controller/src/gas-flows/OracleLayer1GasFeeFlow.test.ts
@@ -50,9 +50,11 @@ const DEFAULT_GAS_PRICE_ORACLE_ADDRESS =
  * @param serializedBuffer - The buffer returned by the serialize method.
  * @returns The mock TypedTransaction object.
  */
-function createMockTypedTransaction(serializedBuffer: Buffer) {
+function createMockTypedTransaction(
+  serializedBuffer: Buffer,
+): jest.Mocked<TypedTransaction> {
   const instance = {
-    serialize: () => serializedBuffer,
+    serialize: (): Buffer => serializedBuffer,
     sign: jest.fn(),
   };
 
@@ -185,6 +187,7 @@ describe('OracleLayer1GasFeeFlow', () => {
         layer1Fee: LAYER_1_FEE_MOCK,
       });
 
+      // eslint-disable-next-line @typescript-eslint/unbound-method
       expect(typedTransactionMock.sign).toHaveBeenCalledTimes(1);
       expect(contractGetOperatorFeeMock).not.toHaveBeenCalled();
     });
@@ -233,6 +236,7 @@ describe('OracleLayer1GasFeeFlow', () => {
       expect(contractMock).toHaveBeenCalledTimes(1);
       const [oracleAddress] = contractMock.mock.calls[0];
       expect(oracleAddress).toBe(DEFAULT_GAS_PRICE_ORACLE_ADDRESS);
+      // eslint-disable-next-line @typescript-eslint/unbound-method
       expect(typedTransactionMock.sign).not.toHaveBeenCalled();
     });