+++ packages/transaction-controller/src/gas-flows/RandomisedEstimationsGasFeeFlow.test.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/unbound-method */
+
 import { toHex } from '@metamask/controller-utils';
 import type EthQuery from '@metamask/eth-query';
 import { GAS_ESTIMATE_TYPES } from '@metamask/gas-fee-controller';